+++ src/app/relationreport/[reportSeq]/components/RelationReportShare.tsx
@@ -8,13 +8,11 @@ import { useParams } from 'next/navigation';
 import useGetRelationReport from '@/apis/useGetRelationReport';
 import { environment } from '../../../../../environments/environment';
 import * as gaEvent from '@/utils/gaEvent';
-import { RelationReportModalContext } from '../page';
+import { useRelationReportContext } from '../context';
 
 const RelationReportShare = () => {
   const params = useParams();
-  const { setIsPreventSharePopupOpen } = React.useContext(
-    RelationReportModalContext,
-  );
+  const { setIsPreventSharePopupOpen } = useRelationReportContext();
   const { data } = useGetRelationReport({
     reportSeq: params.reportSeq as string,
   });
