+++ src/app/relationreport/[reportSeq]/components/RelationReportTitle.tsx
@@ -1,12 +1,12 @@
 import Image from 'next/image';
 import * as React from 'react';
-import { RelationReportModalContext } from '../page';
 import { useParams, useRouter } from 'next/navigation';
 import { t } from '@/utils/translate';
 import useGetRelationReport from '@/apis/useGetRelationReport';
 import * as gaEvent from '@/utils/gaEvent';
 import useGetPlayData from '@/apis/useGetPlayData';
 import webview from '@/utils/webview';
+import { useRelationReportContext } from '../context';
 
 const RelationReportTitle = () => {
   const router = useRouter();
@@ -23,8 +23,8 @@ const RelationReportTitle = () => {
     },
   });
 
-  const { setInviteFriendsPopupOpen, setIsPreventSharePopupOpen } =
-    React.useContext(RelationReportModalContext);
+  const { setInviteFriendsPopupOpen, setEditMoimPopupInfo } =
+    useRelationReportContext();
 
   const createNewMoimButtonClick = () => {
     gaEvent.touchRelationCreateNew({
@@ -41,8 +41,6 @@ const RelationReportTitle = () => {
 
     router.push('/relationreport/create/' + data?.skillSeq);
   };
-
-  const { setEditMoimPopupInfo } = React.useContext(RelationReportModalContext);
 
   const handleKakaoTalkButtonClick = () => {
     setInviteFriendsPopupOpen(true);
