+++ src/app/relationreport/[reportSeq]/components/popup/InviteFriendsPopup.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 import Image from 'next/image';
 import CommonPopup from '@/components/CommonPopup';
 import shareWithKakao from '@/utils/shareWithKakao';
-import { RelationReportModalContext } from '../../page';
 import { copyToClipboard } from '@/utils';
 import { toast } from 'react-toastify';
 import { t } from '@/utils/translate';
@@ -10,6 +9,7 @@ import { useParams } from 'next/navigation';
 import useGetRelationReport from '@/apis/useGetRelationReport';
 import { environment } from '../../../../../../environments/environment';
 import * as gaEvent from '@/utils/gaEvent';
+import { useRelationReportContext } from '../../context';
 
 interface IInviteFriendsPopup {
   onClose: () => void;
@@ -20,9 +20,7 @@ const InviteFriendsPopup = ({ onClose }: IInviteFriendsPopup) => {
   const { data } = useGetRelationReport({
     reportSeq: params.reportSeq as string,
   });
-  const { isInviteFriendsPopupOpen } = React.useContext(
-    RelationReportModalContext,
-  );
+  const { isInviteFriendsPopupOpen } = useRelationReportContext();
 
   const handleCopyLinkButtonClick = () => {
     gaEvent.touchRelationLinkCopy();
